// src/types/user.ts
export interface RegisteredTo {
  first_name: string;
  last_name: string;
  email: string;
  phone: string;
  date_of_birth: string;
  sex: string;
  registration_date: string;
}

export interface PatientInfo {
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  email: string;
  sex: string;
  phone: string;
} 